simplify: extract IPC processing into src/ipc.ts
Move startIpcWatcher() and processTaskIpc() (~350 lines) from index.ts
into a dedicated ipc.ts module. Uses dependency injection via an
IpcDependencies interface to avoid circular imports.

index.ts goes from 867 lines to 515 lines with no behavior change.
